Split monomorphize substitution helpers
