+++ src/typechecker/mod.rs
@@ -30,7 +30,9 @@ mod import_roots;
 mod monomorphize;
 mod monomorphize_dependencies;
 mod monomorphize_inference;
+mod monomorphize_names;
 mod monomorphize_specialized_types;
+mod monomorphize_substitution;
 mod monomorphize_types;
 mod patterns;
 mod program_checking;
